Adapt spillover matrix

This helper function adapts the columns of a provided spillover matrix such that it is compatible with data having the column names provided.

adaptSpillmat(input_sm, out_chs, ...)

# S4 method for matrix,vector
adaptSpillmat(input_sm, out_chs,
  isotope_list = CATALYST::isotope_list)

Arguments

input_sm
a previously calculated spillover matrix.

out_chs
the column names that the prepared output spillover matrix should have. Numeric names as well as names of the form MetalMass(Di), e.g. Ir191Di or Ir191, will be interpreted as masses with associated metals.

...
optional arguments.

isotope_list
named list. Used to validate the input spillover matrix. Names should be metals; list elements numeric vectors of their isotopes. See isotope_list for the list of isotopes used by default.

Value

An adapted spillover matrix with column and row names according to out_chs.

Details

The rules how the spillover matrix is adapted are explained in compCytof.

Examples

# get single-stained control samples
data(ss_exp)
# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)
# debarcode
re <- assignPrelim(x = ss_exp, y = bc_ms)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
# estimate spillover matrix and adapt it
sm <- computeSpillmat(x = re)
chs <- flowCore::colnames(ss_exp)
adaptSpillmat(sm, chs)

#> WARNING: Compensation is likely to be inaccurate.
#>          Spill values for the following interactions
#>          have not been estimated:

#> Ba138Di -> La139Di, Sm154Di

#> Ce140Di -> Ba138Di, La139Di, Pr141Di, Nd142Di, Gd156Di

#> Gd157Di -> Sm152Di, Sm154Di, Gd155Di, Gd156Di, Gd158Di, Gd160Di, Yb173Di

#>         Ba138Di La139Di Ce140Di Pr141Di      Nd142Di     Nd143Di     Nd144Di
#> Ba138Di       1       0       0       0 0.0000000000 0.000000000 0.000000000
#> La139Di       0       1       0       0 0.0000000000 0.000000000 0.000000000
#> Ce140Di       0       0       1       0 0.0000000000 0.000000000 0.000000000
#> Pr141Di       0       0       0       1 0.0000000000 0.000000000 0.000000000
#> Nd142Di       0       0       0       0 1.0000000000 0.002808940 0.002129988
#> Nd143Di       0       0       0       0 0.0071873477 1.000000000 0.019541981
#> Nd144Di       0       0       0       0 0.0015914046 0.001673608 1.000000000
#> Nd145Di       0       0       0       0 0.0004118663 0.000000000 0.009585120
#> Nd146Di       0       0       0       0 0.0000000000 0.000000000 0.001092301
#> Sm147Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Nd148Di       0       0       0       0 0.0116844813 0.006835978 0.013832441
#> Sm149Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Nd150Di       0       0       0       0 0.0014027195 0.000000000 0.003712513
#> Eu151Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Sm152Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Eu153Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Sm154Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Gd155Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Gd156Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Gd157Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Gd158Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Tb159Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Gd160Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Dy161Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#> Dy162Di       0       0       0       0 0.0000000000 0.000000000 0.000000000
#>              Nd145Di      Nd146Di      Sm147Di      Nd148Di      Sm149Di
#> Ba138Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> La139Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Ce140Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Pr141Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Nd142Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Nd143Di 0.0006470948 0.0005555652 0.0000000000 0.0000000000 0.0000000000
#> Nd144Di 0.0161924880 0.0043906441 0.0000000000 0.0000000000 0.0000000000
#> Nd145Di 1.0000000000 0.0382795637 0.0000000000 0.0000000000 0.0000000000
#> Nd146Di 0.0011561573 1.0000000000 0.0000000000 0.0010190798 0.0000000000
#> Sm147Di 0.0000000000 0.0000000000 1.0000000000 0.0223466302 0.0042086772
#> Nd148Di 0.0048295891 0.0156506623 0.0000000000 1.0000000000 0.0000000000
#> Sm149Di 0.0000000000 0.0000000000 0.0006617111 0.0067199238 1.0000000000
#> Nd150Di 0.0000000000 0.0014553347 0.0000000000 0.0012354682 0.0000000000
#> Eu151Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Sm152Di 0.0000000000 0.0000000000 0.0000000000 0.0006252591 0.0003784214
#> Eu153Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Sm154Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Gd155Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Gd156Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Gd157Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Gd158Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Tb159Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Gd160Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Dy161Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#> Dy162Di 0.0000000000 0.0000000000 0.0000000000 0.0000000000 0.0000000000
#>              Nd150Di     Eu151Di     Sm152Di     Eu153Di     Sm154Di
#> Ba138Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> La139Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Ce140Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Pr141Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd142Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd143Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd144Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd145Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd146Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Sm147Di 0.0003632255 0.000000000 0.002890171 0.000000000 0.001394020
#> Nd148Di 0.0132905073 0.000000000 0.000000000 0.000000000 0.000000000
#> Sm149Di 0.0150533552 0.000000000 0.003096888 0.000000000 0.001550511
#> Nd150Di 1.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Eu151Di 0.0000000000 1.000000000 0.000000000 0.007485802 0.000000000
#> Sm152Di 0.0005810608 0.000000000 1.000000000 0.000000000 0.005787269
#> Eu153Di 0.0000000000 0.005576755 0.000000000 1.000000000 0.000000000
#> Sm154Di 0.0000000000 0.000000000 0.005284287 0.000000000 1.000000000
#> Gd155Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd156Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd157Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd158Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Tb159Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd160Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Dy161Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#> Dy162Di 0.0000000000 0.000000000 0.000000000 0.000000000 0.000000000
#>             Gd155Di      Gd156Di      Gd157Di     Gd158Di    Tb159Di
#> Ba138Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> La139Di 0.012757375 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Ce140Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Pr141Di 0.000000000 0.0000000000 0.0174079974 0.000000000 0.00000000
#> Nd142Di 0.000000000 0.0000000000 0.0000000000 0.016088044 0.00000000
#> Nd143Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.01568712
#> Nd144Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Nd145Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Nd146Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Sm147Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Nd148Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Sm149Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Nd150Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Eu151Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Sm152Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Eu153Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Sm154Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Gd155Di 1.000000000 0.0009091198 0.0002479442 0.000000000 0.00000000
#> Gd156Di 0.008035424 1.0000000000 0.0275129168 0.007237057 0.00000000
#> Gd157Di 0.000000000 0.0000000000 1.0000000000 0.000000000 0.00000000
#> Gd158Di 0.000000000 0.0011475542 0.0089223774 1.000000000 0.00000000
#> Tb159Di 0.000000000 0.0000000000 0.0000000000 0.000000000 1.00000000
#> Gd160Di 0.000000000 0.0004098026 0.0006068949 0.005499109 0.00000000
#> Dy161Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#> Dy162Di 0.000000000 0.0000000000 0.0000000000 0.000000000 0.00000000
#>              Gd160Di     Dy161Di     Dy162Di     Dy163Di     Dy164Di
#> Ba138Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> La139Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Ce140Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Pr141Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd142Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd143Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd144Di 1.812179e-02 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd145Di 0.000000e+00 0.016821982 0.000000000 0.000000000 0.000000000
#> Nd146Di 0.000000e+00 0.000000000 0.017519753 0.000000000 0.000000000
#> Sm147Di 0.000000e+00 0.000000000 0.000000000 0.004415254 0.000000000
#> Nd148Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.016181953
#> Sm149Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Nd150Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Eu151Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Sm152Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Eu153Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Sm154Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd155Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd156Di 1.307613e-03 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd157Di 0.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd158Di 5.387737e-03 0.000000000 0.000000000 0.000000000 0.000000000
#> Tb159Di 4.365299e-05 0.000000000 0.000000000 0.000000000 0.000000000
#> Gd160Di 1.000000e+00 0.000000000 0.000000000 0.000000000 0.000000000
#> Dy161Di 1.558121e-03 1.000000000 0.027319642 0.006493813 0.002555605
#> Dy162Di 0.000000e+00 0.014491906 1.000000000 0.040663609 0.006143263
#>              Ho165Di      Er166Di     Er167Di      Er168Di      Tm169Di
#> Ba138Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> La139Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Ce140Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Pr141Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd142Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd143Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd144Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd145Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd146Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Sm147Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd148Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Sm149Di 0.0034629351 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Nd150Di 0.0000000000 0.0123486341 0.000000000 0.000000e+00 0.000000e+00
#> Eu151Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Sm152Di 0.0000000000 0.0000000000 0.000000000 2.222327e-03 0.000000e+00
#> Eu153Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Sm154Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Gd155Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Gd156Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Gd157Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Gd158Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Tb159Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Gd160Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Dy161Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#> Dy162Di 0.0000000000 0.0000000000 0.000000000 0.000000e+00 0.000000e+00
#>              Er170Di      Yb171Di     Yb172Di     Yb173Di      Yb174Di
#> Ba138Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> La139Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Ce140Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Pr141Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd142Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd143Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd144Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd145Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd146Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Sm147Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd148Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Sm149Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Nd150Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Eu151Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Sm152Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Eu153Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Sm154Di 1.379932e-03 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Gd155Di 0.000000e+00 0.0037442465 0.000000000 0.000000000 0.000000e+00
#> Gd156Di 0.000000e+00 0.0000000000 0.003755308 0.000000000 0.000000e+00
#> Gd157Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Gd158Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 3.610301e-03
#> Tb159Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Gd160Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Dy161Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#> Dy162Di 0.000000e+00 0.0000000000 0.000000000 0.000000000 0.000000e+00
#>              Lu175Di      Yb176Di
#> Ba138Di 0.0000000000 0.000000e+00
#> La139Di 0.0000000000 0.000000e+00
#> Ce140Di 0.0000000000 0.000000e+00
#> Pr141Di 0.0000000000 0.000000e+00
#> Nd142Di 0.0000000000 0.000000e+00
#> Nd143Di 0.0000000000 0.000000e+00
#> Nd144Di 0.0000000000 0.000000e+00
#> Nd145Di 0.0000000000 0.000000e+00
#> Nd146Di 0.0000000000 0.000000e+00
#> Sm147Di 0.0000000000 0.000000e+00
#> Nd148Di 0.0000000000 0.000000e+00
#> Sm149Di 0.0000000000 0.000000e+00
#> Nd150Di 0.0000000000 0.000000e+00
#> Eu151Di 0.0000000000 0.000000e+00
#> Sm152Di 0.0000000000 0.000000e+00
#> Eu153Di 0.0000000000 0.000000e+00
#> Sm154Di 0.0000000000 0.000000e+00
#> Gd155Di 0.0000000000 0.000000e+00
#> Gd156Di 0.0000000000 0.000000e+00
#> Gd157Di 0.0000000000 0.000000e+00
#> Gd158Di 0.0000000000 0.000000e+00
#> Tb159Di 0.0030393913 0.000000e+00
#> Gd160Di 0.0000000000 2.873556e-03
#> Dy161Di 0.0000000000 0.000000e+00
#> Dy162Di 0.0000000000 0.000000e+00
#>  [ reached getOption("max.print") -- omitted 14 rows ]