Single-cell debarcoding (2)

Applies separation and mahalanobies distance cutoffs.

applyCutoffs(x, ...)

# S4 method for dbFrame
applyCutoffs(x, mhl_cutoff = 30, sep_cutoffs = NULL)

Arguments

x
a dbFrame.

...
optional arguments.

mhl_cutoff
mahalanobis distance threshold above which events should be unassigned. This argument will be ignored if the mhl_cutoff slot of the input dbFrame is specified.

sep_cutoffs
non-negative numeric of length one or of same length as the number of rows in the bc_key(x). Specifies the distance separation cutoffs between positive and negative barcode populations below which events should be unassigned. If NULL (default), applyCutoffs will try to access the sep_cutoffs slot of the input dbFrame.

Value

Will update the bc_ids and, if not already specified, sep_cutoffs & mhl_cutoff slots of x.

References

Zunder, E.R. et al. (2015). Palladium-based mass tag cell barcoding with a doublet-filtering scheme and single-cell deconvolution algorithm. Nature Protocols 10, 316-333.

Examples

data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

# use global separation cutoff
applyCutoffs(x = re, sep_cutoffs = 0.4)

#> dbFrame object with
#> 20000 events, 6 observables and 20 barcodes:
#> 
#> Current assignments:
#>       6824 event(s) unassigned                                                                             
#> ID    B5  D4  D5  D2  C5  D1  B2  C3  B4  D3  C4  A4  B3  C2  A5  A3  B1  A1 
#> Count 724 718 710 709 705 703 702 695 684 681 663 652 646 638 626 613 599 576
#>              
#> ID    C1  A2 
#> Count 575 557
#> 
#> Separation cutoffs:                                                                              
#> ID     B5  D4  D5  D2  C5  D1  B2  C3  B4  D3  C4  A4  B3  C2  A5  A3  B1  A1 
#> Cutoff 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4 0.4
#>               
#> ID     C1  A2 
#> Cutoff 0.4 0.4
#> 
#> Yields upon debarcoding:
#>       67.66% overall yield                                                                           
#> ID    B5    D4    D5    D2  C5    D1    B2    C3    B4    D3    C4    A4   
#> Yield 74.4% 73.6% 72.9% 73% 72.3% 72.2% 72.1% 71.2% 70.3% 70.1% 67.8% 67.2%
#>                                                      
#> ID    B3    C2    A5    A3    B1    A1    C1    A2   
#> Yield 66.5% 65.6% 64.5% 62.7% 61.5% 58.7% 59.2% 57.3%

# estimate population-specific cutoffs
re <- estCutoffs(x = re)
applyCutoffs(x = re)

#> dbFrame object with
#> 20000 events, 6 observables and 20 barcodes:
#> 
#> Current assignments:
#>       6528 event(s) unassigned                                                                             
#> ID    C5  D1  D4  B5  D2  B2  D5  C3  C4  B4  D3  A4  B3  C2  A5  A3  B1  A1 
#> Count 756 755 751 737 718 714 714 702 697 691 687 676 663 648 633 625 601 576
#>              
#> ID    A2  C1 
#> Count 564 564
#> 
#> Separation cutoffs:                                                                            
#> ID     C5   D1   D4   B5   D2   B2   D5   C3   C4   B4   D3   A4   B3   C2  
#> Cutoff 0.27 0.27 0.32 0.33 0.37 0.35 0.38 0.37 0.29 0.36 0.38 0.32 0.36 0.38
#>                                     
#> ID     A5   A3   B1   A1   A2   C1  
#> Cutoff 0.36 0.35 0.39 0.41 0.34 0.42
#> 
#> Yields upon debarcoding:
#>       69.36% overall yield                                                                               
#> ID    C5  D1    D4    B5  D2    B2    D5    C3    C4    B4    D3    A4    B3   
#> Yield 78% 77.6% 77.1% 76% 74.1% 73.6% 73.3% 71.9% 72.4% 71.2% 70.8% 69.4% 68.1%
#>                                                
#> ID    C2    A5    A3    B1    A1    A2    C1   
#> Yield 66.7% 65.6% 64.3% 61.7% 58.7% 58.4% 58.3%