Single-cell debarcoding (1)

Assigns a preliminary barcode ID to each event.

assignPrelim(x, y, ...)

# S4 method for flowFrame,data.frame
assignPrelim(x, y, cofactor = 10,
  verbose = TRUE)

# S4 method for flowFrame,vector
assignPrelim(x, y, cofactor = 10,
  verbose = TRUE)

# S4 method for character,data.frame
assignPrelim(x, y, cofactor = 10,
  verbose = TRUE)

# S4 method for character,vector
assignPrelim(x, y, cofactor = 10,
  verbose = TRUE)

Arguments

x
a flowFrame or character of an FCS file name.

y
the debarcoding scheme. A binary matrix with sample names as row names and numeric masses as column names OR a vector of numeric masses corresponding to barcode channels. When the latter is supplied, assignPrelim will create a scheme of the appropriate format internally.

...
optional arguments.

cofactor
numeric. Cofactor used for asinh transformation.

verbose
logical. Should extra information on progress be reported?

Value

Returns a dbFrame containing measurement intensities, the debarcoding key, a numeric verctor of barcode IDs and separations between positive and negative barcode populations, and barcode intensities normalized by population.

References

Zunder, E.R. et al. (2015). Palladium-based mass tag cell barcoding with a doublet-filtering scheme and single-cell deconvolution algorithm. Nature Protocols 10, 316-333.

Examples

data(sample_ff, sample_key)
assignPrelim(x = sample_ff, y = sample_key)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

#> dbFrame object with
#> 20000 events, 6 observables and 20 barcodes:
#> 
#> Current assignments:
#>       0 event(s) unassigned                                                                           
#> ID    A1   A2   A3   A4   A5   B1   B2   B3   B4   B5   C1   C2   C3   C4  
#> Count 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000 1000
#>                                    
#> ID    C5   D1   D2   D3   D4   D5  
#> Count 1000 1000 1000 1000 1000 1000