FlowSOM clustering & ConsensusClusterPlus metaclustering

cluster will first group cells into xdimxydim clusters using FlowSOM, and subsequently perform metaclustering with ConsensusClusterPlus into 2 through maxK clusters. In the returned daFrame, those antigens used for clustering will be labelled as 'type' markers, and the remainder of antigens as 'state' markers.

cluster(x, ...)

# S4 method for daFrame
cluster(x, cols_to_use, xdim = 10, ydim = 10,
  maxK = 20, verbose = TRUE, seed = 1)

Arguments

x
a daFrame.

...
optional arguments.

cols_to_use
a character vector. Specifies which antigens to use for clustering.

xdim, ydim
numeric. Specify the grid size of the self-orginizing map. The default 10x10 grid will yield 100 clusters.

maxK
numeric. Specifies the maximum number of clusters to evaluate in the metaclustering. For maxK = 20, for example, metaclustering will be performed for 2 through 20 clusters.

verbose
logical. Should information on progress be reported?

seed
numeric. Sets random seed in ConsensusClusterPlus().

Value

The function will add information to the following slots of the input daFrame (and return it):

rowData

cluster_id: each cell's cluster ID as inferred by FlowSOM. One of 1, ..., xdimxydim.

colData

marker_class: "type" or "state". Specifies whether an antigen has been used for clustering or not, respectively.

metadata

SOM_codes: a table with dimensions K x (# cell type markers), where K = xdim x ydim. Contains the SOM codes.

cluster_codes: a table with dimensions K x (maxK + 1). Contains the cluster codes for all metaclustering.

delta_area: a ggplot object. See above for details.

Details

The delta area represents the amount of extra cluster stability gained when clustering into k groups as compared to k-1 groups. It can be expected that high stability of clusters can be reached when clustering into the number of groups that best fits the data. The "natural" number of clusters present in the data should thus corresponds to the value of k where there is no longer a considerable increase in stability (pleateau onset).

References

Nowicka M, Krieg C, Weber LM et al. CyTOF workflow: Differential discovery in high-throughput high-dimensional cytometry datasets. F1000Research 2017, 6:748 (doi: 10.12688/f1000research.11622.1)

Examples

data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

#> Error in as.list.default(X): no method for coercing this S4 class to a vector

# specify antigens to use for clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33",
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
(re <- cluster(re, cols_to_use=lineage))

#> Error in cluster(re, cols_to_use = lineage): object 're' not found