Compensate CyTOF data

Compensates a mass spectrometry based experiment using a provided spillover matrix & assuming a linear spillover in the experiment.

compCytof(x, y, ...)

# S4 method for flowFrame,matrix
compCytof(x, y, out_path = NULL,
  method = "flow", isotope_list = CATALYST::isotope_list)

# S4 method for flowSet,ANY
compCytof(x, y, out_path = NULL, method = "flow")

# S4 method for character,matrix
compCytof(x, y, out_path = NULL,
  method = "flow")

# S4 method for ANY,data.frame
compCytof(x, y, out_path = NULL, method = "flow")

Arguments

x
a flowFrame OR a character string specifying the location of FCS files that should be compensates.

y
a spillover matrix.

...
optional arguments.

out_path
a character string. If specified, compensated FCS files will be generated in this location. If x is a character string, file names will be inherited from uncompensated FCS files and given extension "_comped".

method
"flow" or "nnls".

isotope_list
named list. Used to validate the input spillover matrix. Names should be metals; list elements numeric vectors of their isotopes. See isotope_list for the list of isotopes used by default.

Value

Compensates the input flowFrame or, if x is a character string, all FCS files in the specified location. If out_path=NULL (the default), returns a flowFrame containing the compensated data. Otherwise, compensated data will be written to the specified location as FCS 3.0 standard files.

Details

If the spillover matrix (SM) does not contain the same set of columns as the input experiment, it will be adapted according to the following rules:

columns present in the SM but not in the input data will be removed from it

non-metal columns present in the input but not in the SM will be added such that they do neither receive nor cause spill

metal columns that have the same mass as a channel present in the SM will receive (but not emit) spillover according to that channel

if an added channel could potentially receive spillover (as it has +/-1M or +16M of, or is of the same metal type as another channel measured), a warning will be issued as there could be spillover interactions that have been missed and may lead to faulty compensation

Examples

# get single-stained control samples
data(ss_exp)

# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)

# debarcode
re <- assignPrelim(x = ss_exp, y = bc_ms)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
spillMat <- computeSpillmat(x = re)
compCytof(x = ss_exp, y = spillMat)

#> WARNING: Compensation is likely to be inaccurate.
#>          Spill values for the following interactions
#>          have not been estimated:

#> Ba138Di -> La139Di, Sm154Di

#> Ce140Di -> Ba138Di, La139Di, Pr141Di, Nd142Di, Gd156Di

#> Gd157Di -> Sm152Di, Sm154Di, Gd155Di, Gd156Di, Gd158Di, Gd160Di, Yb173Di

#> flowFrame object 'anonymous'
#> with 10000 cells and 39 observables:
#>       name          desc       range minRange    maxRange
#> 1  Ba138Di         138Ba  4191.74365        0  4191.74365
#> 2  La139Di     139La-CD8   778.13989        0   778.13989
#> 3  Ce140Di         140Ce   169.58528        0   169.58528
#> 4  Pr141Di    141Pr-CD64  1277.42114        0  1277.42114
#> 5  Nd142Di    142Nd-CD23  1654.36560        0  1654.36560
#> 6  Nd143Di    143Nd-CD68  2198.46680        0  2198.46680
#> 7  Nd144Di    144Nd-CD36  2547.50610        0  2547.50610
#> 8  Nd145Di     145Nd-CD4  1968.55127        0  1968.55127
#> 9  Nd146Di    146Nd-CD68  3003.05933        0  3003.05933
#> 10 Sm147Di     147Sm-CD3  3166.04346        0  3166.04346
#> 11 Nd148Di    148Nd-CD20  2323.83203        0  2323.83203
#> 12 Sm149Di    149Sm-CD20  3839.22925        0  3839.22925
#> 13 Nd150Di    150Nd-CD68  2266.47339        0  2266.47339
#> 14 Eu151Di   151Eu-CD123  6530.27783        0  6530.27783
#> 15 Sm152Di    152Sm-CD99  4116.97168        0  4116.97168
#> 16 Eu153Di    153Eu-CD68  3351.10181        0  3351.10181
#> 17 Sm154Di    154Sm-CD15  1913.62158        0  1913.62158
#> 18 Gd155Di   155Gd-CD273  8811.60156        0  8811.60156
#> 19 Gd156Di    156Gd-CD93  3172.47607        0  3172.47607
#> 20 Gd157Di         157Gd    96.05315        0    96.05315
#> 21 Gd158Di    158Gd-CD15  2703.31030        0  2703.31030
#> 22 Tb159Di   159Tb-CD192  4230.27100        0  4230.27100
#> 23 Gd160Di    160Gd-CD45  4405.06299        0  4405.06299
#> 24 Dy161Di 161Dy-CD66ace 10547.04883        0 10547.04883
#> 25 Dy162Di   162Dy-CXCR4  4564.75635        0  4564.75635
#> 26 Dy163Di    163Dy-CD22  5905.81445        0  5905.81445
#> 27 Dy164Di     164Dy-CD7  3505.41919        0  3505.41919
#> 28 Ho165Di     165Ho-CD4  7356.32715        0  7356.32715
#> 29 Er166Di    166Er-CD32  4321.32178        0  4321.32178
#> 30 Er167Di    167Er-CD16  4528.40234        0  4528.40234
#> 31 Er168Di    168Er-CD14   769.98413        0   769.98413
#> 32 Tm169Di    169Tm-CD99  4413.05371        0  4413.05371
#> 33 Er170Di     170Er-CD7  3510.09277        0  3510.09277
#> 34 Yb171Di  171Yb-HLA.DR  4580.26221        0  4580.26221
#> 35 Yb172Di 172Yb-HLA.ABC  3021.48706        0  3021.48706
#> 36 Yb173Di     173Yb-CD3  7392.21143        0  7392.21143
#> 37 Yb174Di    174Yb-CD8b  2360.97510        0  2360.97510
#> 38 Lu175Di  175Lu-HLA.DR  5659.57666        0  5659.57666
#> 39 Yb176Di    176Yb-CD45  5700.28271        0  5700.28271
#> 0 keywords are stored in the 'description' slot