Compute spillover matrix

Computes a spillover matrix from a priori identified single-positive populations.

computeSpillmat(x, ...)

# S4 method for dbFrame
computeSpillmat(x, method = "default",
  interactions = "default", trim = 0.5, th = 1e-05)

Arguments

x
a dbFrame.

...
optional arguments.

method
"default" or "classic". Specifies the function to be used for spillover estimation (see below for details).

interactions
"default" or "all". Specifies which interactions spillover should be estimated for. The default exclusively takes into consideration interactions that are sensible from a chemical and physical point of view (see below for more details).

trim
numeric. Specifies the trim value used for estimation of spill values. Note that trim = 0.5 is equivalent to using medians.

th
single non-negative numeric. Specifies the threshold value below which spill estimates will be set to 0.

Value

Returns a square compensation matrix with dimensions and dimension names matching those of the input flowFrame. Spillover is assumed to be linear, and, on the basis of their additive nature, spillover values are computed independently for each interacting pair of channels.

Details

The default method estimates the spillover as the median ratio between the unstained spillover receiving and the stained spillover emitting channel in the corresponding single stained populations.

method = "classic" will compute the slope of a line through the medians (or trimmed means) of stained and unstained populations. The medians (or trimmed means) computed from events that are i) negative in the respective channels; and, ii) not assigned to interacting channels; and, iii) not unassigned are subtracted as to account for background.

interactions="default" considers only expected interactions, that is, M+/-1 (detection sensitivity), M+16 (oxide formation) and channels measuring metals that are potentially contaminated by isotopic impurites (see reference below and isotope_list).

interaction="all" will estimate spill for all n x n - n interactions, where n denotes the number of single-color controls (= nrow(bc_key(re))).

References

Coursey, J.S., Schwab, D.J., Tsai, J.J., Dragoset, R.A. (2015). Atomic weights and isotopic compositions, (available at http://physics.nist.gov/Comp).

Examples

# get single-stained control samples
data(ss_exp)
# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)
# debarcode single-positive populations
re <- assignPrelim(x = ss_exp, y = bc_ms)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
head(computeSpillmat(x = re))

#>         Ba138Di La139Di Ce140Di Pr141Di      Nd142Di     Nd143Di     Nd144Di
#> La139Di       0       1       0       0 0.0000000000 0.000000000 0.000000000
#> Pr141Di       0       0       0       1 0.0000000000 0.000000000 0.000000000
#> Nd142Di       0       0       0       0 1.0000000000 0.002808940 0.002129988
#> Nd143Di       0       0       0       0 0.0071873477 1.000000000 0.019541981
#> Nd144Di       0       0       0       0 0.0015914046 0.001673608 1.000000000
#> Nd145Di       0       0       0       0 0.0004118663 0.000000000 0.009585120
#>              Nd145Di      Nd146Di Sm147Di Nd148Di Sm149Di Nd150Di Eu151Di
#> La139Di 0.0000000000 0.0000000000       0       0       0       0       0
#> Pr141Di 0.0000000000 0.0000000000       0       0       0       0       0
#> Nd142Di 0.0000000000 0.0000000000       0       0       0       0       0
#> Nd143Di 0.0006470948 0.0005555652       0       0       0       0       0
#> Nd144Di 0.0161924880 0.0043906441       0       0       0       0       0
#> Nd145Di 1.0000000000 0.0382795637       0       0       0       0       0
#>         Sm152Di Eu153Di Sm154Di    Gd155Di Gd156Di  Gd157Di    Gd158Di
#> La139Di       0       0       0 0.01275737       0 0.000000 0.00000000
#> Pr141Di       0       0       0 0.00000000       0 0.017408 0.00000000
#> Nd142Di       0       0       0 0.00000000       0 0.000000 0.01608804
#> Nd143Di       0       0       0 0.00000000       0 0.000000 0.00000000
#> Nd144Di       0       0       0 0.00000000       0 0.000000 0.00000000
#> Nd145Di       0       0       0 0.00000000       0 0.000000 0.00000000
#>            Tb159Di    Gd160Di    Dy161Di Dy162Di Dy163Di Dy164Di Ho165Di
#> La139Di 0.00000000 0.00000000 0.00000000       0       0       0       0
#> Pr141Di 0.00000000 0.00000000 0.00000000       0       0       0       0
#> Nd142Di 0.00000000 0.00000000 0.00000000       0       0       0       0
#> Nd143Di 0.01568712 0.00000000 0.00000000       0       0       0       0
#> Nd144Di 0.00000000 0.01812179 0.00000000       0       0       0       0
#> Nd145Di 0.00000000 0.00000000 0.01682198       0       0       0       0
#>         Er166Di Er167Di Er168Di Tm169Di Er170Di Yb171Di Yb172Di Yb173Di Yb174Di
#> La139Di       0       0       0       0       0       0       0       0       0
#> Pr141Di       0       0       0       0       0       0       0       0       0
#> Nd142Di       0       0       0       0       0       0       0       0       0
#> Nd143Di       0       0       0       0       0       0       0       0       0
#> Nd144Di       0       0       0       0       0       0       0       0       0
#> Nd145Di       0       0       0       0       0       0       0       0       0
#>         Lu175Di Yb176Di
#> La139Di       0       0
#> Pr141Di       0       0
#> Nd142Di       0       0
#> Nd143Di       0       0
#> Nd144Di       0       0
#> Nd145Di       0       0