FCS file Concatenation

Concatenates all input data to a single file or object.

concatFCS(x, ...)

# S4 method for flowSet
concatFCS(x, out_path = NULL, by_time = TRUE,
  file_num = FALSE, pars = NULL, desc = NULL)

# S4 method for character
concatFCS(x, out_path = NULL, by_time = TRUE,
  file_num = FALSE)

# S4 method for list
concatFCS(x, out_path = NULL, by_time = TRUE,
  file_num = FALSE)

Arguments

x
can be either a flowSet, a list of flowFrames, a character specifying the location of the FCS files to be concatinated, or a vector of FCS file names.

...
optional arguments.

out_path
character string. If specified, an FCS file of the concatenated data will be written to this location. If NULL (default), a flowFrame will be returned.

by_time
logical. Specifies whether files should be ordered by time of acquisition.

file_num
logical. Specifies whether a file number column should be added.

pars, desc
optional character vectors of channel names & descriptions to use when merging files.

Value

a flowFrame containing measurement intensities of all input data or a character of the FCS file name.

Examples

data(raw_data)
concatFCS(raw_data)

#> flowFrame object 'raw_data_concat.fcs'
#> with 5000 cells and 67 observables:
#>              name  desc   range minRange maxRange
#> $P1          Time  $P1S 1022077        0  1022076
#> $P2  Event_length  $P2S      39        0       38
#> $P3         Y89Di  $P3S   11697        0    11696
#> $P4        Ru96Di  $P4S     663        0      662
#> $P5        Ru98Di  $P5S    1049        0     1048
#> $P6        Ru99Di  $P6S     281        0      280
#> $P7       Ru100Di  $P7S     556        0      555
#> $P8       Ru101Di  $P8S      88        0       87
#> $P9       Ru102Di  $P9S     161        0      160
#> $P10      Rh103Di $P10S     636        0      635
#> $P11      Ru104Di $P11S     161        0      160
#> $P12      Pd105Di $P12S    1518        0     1517
#> $P13      Pd106Di $P13S    8490        0     8489
#> $P14      Pd108Di $P14S    1773        0     1772
#> $P15      Pd110Di $P15S    1465        0     1464
#> $P16      In113Di $P16S    2053        0     2052
#> $P17      In115Di $P17S    1986        0     1985
#> $P18      Sn120Di $P18S    6784        0     6783
#> $P19      Ba138Di $P19S   14014        0    14013
#> $P20      La139Di $P20S     183        0      182
#> $P21      Ce140Di $P21S    3849        0     3848
#> $P22      Pr141Di $P22S     632        0      631
#> $P23      Nd142Di $P23S   26648        0    26647
#> $P24      Nd143Di $P24S    4152        0     4151
#> $P25      Nd144Di $P25S    2378        0     2377
#> $P26      Nd145Di $P26S    2728        0     2727
#> $P27      Nd146Di $P27S    4247        0     4246
#> $P28      Sm147Di $P28S     973        0      972
#> $P29      Nd148Di $P29S     279        0      278
#> $P30      Sm149Di $P30S     444        0      443
#> $P31      Nd150Di $P31S     950        0      949
#> $P32      Eu151Di $P32S    2752        0     2751
#> $P33      Sm152Di $P33S     188        0      187
#> $P34      Eu153Di $P34S    3418        0     3417
#> $P35      Sm154Di $P35S     283        0      282
#> $P36      Gd155Di $P36S     281        0      280
#> $P37      Gd156Di $P37S     441        0      440
#> $P38      Gd157Di $P38S     270        0      269
#> $P39      Gd158Di $P39S    9378        0     9377
#> $P40      Tb159Di $P40S    1725        0     1724
#> $P41      Gd160Di $P41S    2409        0     2408
#> $P42      Dy161Di $P42S     177        0      176
#> $P43      Dy162Di $P43S     902        0      901
#> $P44      Dy163Di $P44S     179        0      178
#> $P45      Dy164Di $P45S    1057        0     1056
#> $P46      Ho165Di $P46S    3946        0     3945
#> $P47      Er166Di $P47S    2784        0     2783
#> $P48      Er167Di $P48S     995        0      994
#> $P49      Er168Di $P49S    1747        0     1746
#> $P50      Tm169Di $P50S     456        0      455
#> $P51      Er170Di $P51S    1463        0     1462
#> $P52      Yb171Di $P52S    2078        0     2077
#> $P53      Yb172Di $P53S     958        0      957
#> $P54      Yb173Di $P54S    1780        0     1779
#> $P55      Yb174Di $P55S    2761        0     2760
#> $P56      Lu175Di $P56S    4049        0     4048
#> $P57      Yb176Di $P57S    4782        0     4781
#> $P58      Ir191Di $P58S    6550        0     6549
#> $P59      Ir193Di $P59S    9450        0     9449
#> $P60      Pt194Di $P60S   18950        0    18949
#> $P61      Pt195Di $P61S   19275        0    19274
#> $P62      Pb208Di $P62S   22849        0    22848
#> $P63      Bi209Di $P63S    2713        0     2712
#> $P64       Center $P64S    2711        0     2710
#> $P65       Offset $P65S     593        0      592
#> $P66        Width $P66S     130        0      129
#> $P67     Residual $P67S     488        0      487
#> 208 keywords are stored in the 'description' slot