Differential analysis frame class

Represents the data returned by and used throughout differential analysis.

assays

list of length one containing the arcsinh-transformed expressions.

rowData

the metadata information for each event, and its cluster ID as inferred by the initial FlowSOM clustering.

colData

a data.frame with the following columns:

marker_name original column name in the input flowSet

marker_class one of "type" or "state"

metadata

a named list containing:

design: the original metadata-table

panel: the original panel-table

n_cells: the number of events measured per sample

SOM_codes: a k x p matrix of SOM codes, where k = no. of clusters, and p = no. of measurement parameters

cluster_codes: cluster codes for the initial FlowSOM clustering, the ConsensusClusterPlus metaclustering, and manual mergings done with mergeClusters

daFrame(x, panel, md, cols_to_use = NULL, cofactor = 5,
  panel_cols = list(channel = "fcs_colname", antigen = "antigen"),
  md_cols = list(file = "file_name", id = "sample_id", factors =
  c("condition", "patient_id")))

Arguments

x
a flowSet holding all samples OR a character vector that specifies a path to set of FCS files.

panel
a 2 column data.frame that contains for each marker of interest i) its column name in the FCS file, and ii) the targeted protein marker.

md
a data.frame with columns describing the experiment. An exemplary metadata table could look as follows:

file_name: the FCS file name

sample_id: a unique sample identifier

condition: brief sample description (e.g. REF)

patient_id: the patient ID

cols_to_use
a logical vector OR numeric vector of indices OR character vector of column names. Specifies which columns to keep from the input flowSet/FCS files.

cofactor
numeric. Cofactor to use for arcsinh-transformation.

panel_cols
a named list specifying column names of panel that contain i) the original channel names in fs, and ii) the targeted protein marker. Elements must be named "channel" and "antigen".

md_cols
a named list specifying column names of md that contain i) the FCS file names, ii) unique sample identifiers, and iii) a character vector of factors descriptive of the samples (e.g. condition, treatment, ect.). Elements must be named "file", "id", and "factors".

Value

an object of class SummarizedExperiment.