Debarcoding frame class

This class represents the data returned by and used throughout debarcoding.

Details

Objects of class dbFrame hold all data required for debarcoding:

as the initial step of single-cell deconcolution, assignPrelim will return a dbFrame containing the input measurement data, barcoding scheme, and preliminary assignments.

assignments will be made final by applyCutoffs. Optionally, population-specific separation cutoffs may be estimated by running estCutoffs prior to this.

plotYields, plotEvents and plotMahal aim to guide devoncolution parameter selection, and to give a sense of the resulting barcode assignment quality.

show(dbFrame) will display

the dimensionality of the measurement data and number of barcodes

current assignments in order of decreasing population size

current separation cutoffs

the mean & per-population yield that'll be achieved upon debarcoding

Slots

exprs

a matrix containing raw intensities of the input flowFrame.

bc_key

binary barcoding scheme with numeric masses as column names and samples names as row names OR a numeric vector of barcode masses.

bc_ids

vector of barcode IDs. If a barcoding scheme is supplied, the respective binary code's row name, else, the mass of the respective barcode channel.

deltas

numeric vector of separations between positive and negative barcode populations computed from normalized barcode intensities.

normed_bcs

matrix containing normalized barcode intensities.

mhl_dists

mahalanobis distances.

sep_cutoffs

numeric vector of distance separation cutoffs between positive and negative barcode populations above which events will be unassigned.

mhl_cutoff

non-negative and non-zero numeric value specifying the Mahalanobis distance below which events will be unassigned.

counts

matrix of dimension (# barcodes)x(101) where each row contains the number of events within a barcode for which positive and negative populations are separated by a distance between in [0,0.01), ..., [0.99,1], respectively.

yields

a matrix of dimension (# barcodes)x(101) where each row contains the percentage of events within a barcode that will be obtained after applying a separation cutoff of 0, 0.01, ..., 1, respectively.