Extraction and replacement methods for objects of class dbFrame

Methods for replacing and accessing slots in a dbFrame.

bc_key(x)

bc_ids(x)

deltas(x)

normed_bcs(x)

mhl_dists(x)

sep_cutoffs(x)

mhl_cutoff(x)

counts(x)

yields(x)

# S4 method for dbFrame
exprs(object)

# S4 method for dbFrame
bc_key(x)

# S4 method for dbFrame
bc_ids(x)

# S4 method for dbFrame
deltas(x)

# S4 method for dbFrame
normed_bcs(x)

# S4 method for dbFrame
mhl_dists(x)

# S4 method for dbFrame
sep_cutoffs(x)

# S4 method for dbFrame
mhl_cutoff(x)

# S4 method for dbFrame
counts(x)

# S4 method for dbFrame
yields(x)

# S4 method for dbFrame,numeric
mhl_cutoff(x) <- value

# S4 method for dbFrame,ANY
mhl_cutoff(x) <- value

# S4 method for dbFrame,numeric
sep_cutoffs(x) <- value

# S4 method for dbFrame,ANY
sep_cutoffs(x) <- value

Arguments

x, object
a dbFrame.

value
the replacement value.

Value

exprs

extracts the raw data intensities.

bc_key

extracts the barcoding scheme.

bc_ids

extracts currently made event assignments.

deltas

extracts barcode separations computed from normalized intensities. sep_cutoffs apply to these values (see applyCutoffs).

normed_bcs

extracts normalized barcode intensities (see assignPrelim).

sep_cutoffs, sep_cutoffs<-

extracts or replaces separation cutoffs. If option sep_cutoffs is not specified, these will be used by applyCutoffs. Replacement value must be a non- negative numeric with length one or same length as the number of barcodes.

mhl_cutoff, mhl_cutoff<-

extracts or replaces the Mahalanobis distance threshold above which events are to be unassigned. Replacement value must be a single non-negative and non-zero numeric.

counts

extract the counts matrix (see dbFrame).

yields

extract the yields matrix (see dbFrame).

Examples

data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

# set global cutoff parameter
sep_cutoffs(re) <- 0.4
re <- applyCutoffs(x = re)

# subset a specific population, e.g. A1: 111000
a1 <- bc_ids(re) == "A1"
head(exprs(sample_ff[a1, ]))

#>      (Pd102)Di (Pd104)Di (Pd105)Di (Pd106)Di  (Pd108)Di  (Pd110)Di
#> [1,]  653.0775  581.6407  537.1787  2.446666 -0.8309277 -0.7016335
#> [2,] 3109.9167 2873.6235 2423.5984 12.111660 -0.9389960 -0.2946386
#> [3,]  937.2463  987.2780  941.2000  1.188081  0.3483154  0.1288555
#> [4,] 1310.6329 1343.3278 1154.9589 17.041691  0.9958168  1.8983954
#> [5,]  966.3564  822.2518  860.4620  6.607563 10.2502708 -0.8406625
#> [6,] 3375.7522 3297.5425 2724.3186 11.865969 -0.4299290 -0.9490275

# subset unassigned events
unassigned <- bc_ids(re) == 0
head(exprs(sample_ff[unassigned, ]))

#>        (Pd102)Di   (Pd104)Di  (Pd105)Di    (Pd106)Di    (Pd108)Di    (Pd110)Di
#> [1,] 1293.755981 1215.938721   15.61643 2086.5397949 2741.3303223  977.9278564
#> [2,] 1894.031738   32.926933 5747.51318 3215.8105469 1339.3760986 2317.1325684
#> [3,]    5.898072   21.270260  630.17377 4146.6396484 3313.8378906 2015.3986816
#> [4,] 2108.750000 2081.191895 1918.94409 3681.1892090    0.3493983 1058.1044922
#> [5,]    2.036946   26.491686   43.31301   -0.6407154   -0.3375399   35.4466972
#> [6,] 1830.069458    1.712605 2432.71558 4222.1181641 3279.1162109   -0.8261126