Estimation of distance separation cutoffs

For each sample, estimates a cutoff parameter for the distance between positive and negative barcode populations.

estCutoffs(x, ...)

# S4 method for dbFrame
estCutoffs(x)

Arguments

x
a dbFrame.

...
optional arguments.

Value

Will update the sep_cutoffs slot of the input dbFrame and return the latter.

Details

For the estimation of cutoff parameters, we considered yields upon debarcoding as a function of the applied cutoffs. Commonly, this function will be characterized by an initial weak decline, where doublets are excluded, and subsequent rapid decline in yields to zero. In between, low numbers of counts with intermediate barcode separation give rise to a plateau. As an adequate cutoff estimate, we target the point that approximately marks the end of the plateau regime and the onset of yield decline. To facilitate robust cutoff estimation, we fit a linear and a three-parameter log-logistic function to the yields function: $$f(x)=\frac{d}{1+e^{b(log(x)-log(e))}}$$ The goodness of the linear fit relative to the log-logistic fit is weighed with: $$w=\frac{RSS_{log-logistic}}{RSS_{log-logistic}+RSS_{linear}}$$ and the cutoffs for both functions are defined as: $$c_{linear}=-\frac{\beta_0}{2\beta_1}$$ $$c_{log-logistic}=argmin_x\{\frac{\vert f'(x)\vert}{ f(x)}>0.1\}$$ The final cutoff estimate is defined as the weighted mean between these estimates: $$c=(1-w)\cdot c_{log-logistic}+w\cdot c_{linear}$$

References

Finney, D.J. (1971). Probit Analsis. Journal of Pharmaceutical Sciences 60, 1432.

Examples

data(sample_ff, sample_key)
# assign preliminary IDs
re <- assignPrelim(x = sample_ff, y = sample_key)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

# estimate separation cutoffs
re <- estCutoffs(x = re)
# view exemplary estimate
plotYields(re, "A1")

#> We recommend that you use the dev version of ggplot2 with `ggplotly()`
#> Install it with: `devtools::install_github('hadley/ggplot2')`

#> [[1]]
#>