Guess parameter panel

Helper function to parse information from the parameters slot of a flowFrame/flowSet.

guessPanel(x, ...)

# S4 method for flowFrame
guessPanel(x)

# S4 method for flowSet
guessPanel(x, index = 1)

Arguments

x
a flowFrame or flowSet.

...
optional arguments.

index
numeric. If x is a flowSet object, this index specifies which flowFrame to extract.

Value

a data.frame with the following columns:

name: the parameter name as extracted from the input flowFrame,

desc: the parameter description as extracted from the input flowFrame,

antigen: the targeted protein markers, and

use_channel: logical. If TRUE, the channel is expected to contain a marker and will be kept.

Examples

# examplary data with Time, DNA, BC channels, etc.
data(raw_data)
guessPanel(raw_data)

#>     fcs_colname         desc desc.1 antigen use_channel
#> 1          Time         <NA>   <NA>    <NA>           0
#> 2  Event_length         <NA>   <NA>    <NA>           0
#> 3         Y89Di      89Y_BC1    89Y     BC1           0
#> 4        Ru96Di    96Ru_Vol1   96Ru    Vol1           1
#> 5        Ru98Di    98Ru_Vol2   98Ru    Vol2           1
#> 6        Ru99Di    99Ru_Vol3   99Ru    Vol3           1
#> 7       Ru100Di   100Ru_Vol4  100Ru    Vol4           1
#> 8       Ru101Di   101Ru_Vol5  101Ru    Vol5           1
#> 9       Ru102Di   102Ru_Vol6  102Ru    Vol6           1
#> 10      Rh103Di    103Rh_BC2  103Rh     BC2           0
#> 11      Ru104Di   104Ru_Vol7  104Ru    Vol7           1
#> 12      Pd105Di    105Pd_BC3  105Pd     BC3           0
#> 13      Pd106Di    106Pd_BC4  106Pd     BC4           0
#> 14      Pd108Di    108Pd_BC5  108Pd     BC5           0
#> 15      Pd110Di    110Pd_BC6  110Pd     BC6           0
#> 16      In113Di    113In_BC7  113In     BC7           0
#> 17      In115Di    115In_BC8  115In     BC8           0
#> 18      Sn120Di        120Sn  120Sn    <NA>           0
#> 19      Ba138Di        138Ba  138Ba    <NA>           0
#> 20      La139Di        139La  139La    <NA>           0
#> 21      Ce140Di  140Ce_Beads  140Ce   Beads           1
#> 22      Pr141Di   141Pr_CD64  141Pr    CD64           1
#> 23      Nd142Di   142Nd_CD38  142Nd    CD38           1
#> 24      Nd143Di   143Nd_CD68  143Nd    CD68           1
#> 25      Nd144Di   144Nd_CD36  144Nd    CD36           1
#> 26      Nd145Di  145Nd_CD197  145Nd   CD197           1
#> 27      Nd146Di 146Nd_CD45RA  146Nd  CD45RA           1
#> 28      Sm147Di  147Sm_CD206  147Sm   CD206           1
#> 29      Nd148Di   148Nd_CD20  148Nd    CD20           1
#> 30      Sm149Di  149Sm_CD204  149Sm   CD204           1
#> 31      Nd150Di  150Nd_CD123  150Nd   CD123           1
#> 32      Eu151Di  151Eu_CD11b  151Eu   CD11b           1
#> 33      Sm152Di  152Sm_FOXP3  152Sm   FOXP3           1
#> 34      Eu153Di 153Eu_Slamf7  153Eu  Slamf7           1
#> 35      Sm154Di 154Sm_CTLA-4  154Sm  CTLA-4           1
#> 36      Gd155Di  155Gd_CD278  155Gd   CD278           1
#> 37      Gd156Di   156Gd_CD93  156Gd    CD93           1
#> 38      Gd157Di        157Gd  157Gd    <NA>           0
#> 39      Gd158Di   158Gd_CD15  158Gd    CD15           1
#> 40      Tb159Di  159Tb_CD192  159Tb   CD192           1
#> 41      Gd160Di  160Gd_CD274  160Gd   CD274           1
#> 42      Dy161Di  161Dy_Tim-3  161Dy   Tim-3           1
#> 43      Dy162Di   162Dy_CD25  162Dy    CD25           1
#> 44      Dy163Di   163Dy_PD-1  163Dy    PD-1           1
#> 45      Dy164Di    164Dy_CD7  164Dy     CD7           1
#> 46      Ho165Di    165Ho_CD4  165Ho     CD4           1
#> 47      Er166Di   166Er_CD32  166Er    CD32           1
#> 48      Er167Di   167Er_CD16  167Er    CD16           1
#> 49      Er168Di   168Er_CD14  168Er    CD14           1
#> 50      Tm169Di  169Tm_CD163  169Tm   CD163           1
#> 51      Er170Di  170Er_CD169  170Er   CD169           1
#> 52      Yb171Di   171Yb_CD86  171Yb    CD86           1
#> 53      Yb172Di 172Yb_HLA-DR  172Yb  HLA-DR           1
#> 54      Yb173Di    173Yb_CD3  173Yb     CD3           1
#> 55      Yb174Di  174Yb_CD11c  174Yb   CD11c           1
#> 56      Lu175Di   175Lu_CD8a  175Lu    CD8a           1
#> 57      Yb176Di   176Yb_CD45  176Yb    CD45           1
#> 58      Ir191Di   191Ir_DNA1  191Ir    DNA1           0
#> 59      Ir193Di   193Ir_DNA2  193Ir    DNA2           0
#> 60      Pt194Di 194Pt_cisPt1  194Pt  cisPt1           1
#> 61      Pt195Di 195Pt_cisPt2  195Pt  cisPt2           1
#> 62      Pb208Di        208Pb  208Pb    <NA>           0
#> 63      Bi209Di    209Bi_BC9  209Bi     BC9           0
#> 64       Center         <NA>   <NA>    <NA>           0
#> 65       Offset         <NA>   <NA>    <NA>           0
#> 66        Width         <NA>   <NA>    <NA>           0
#> 67     Residual         <NA>   <NA>    <NA>           0