Manual cluster merging

mergeClusters provides a simple wrapper to store a manual merging inside the input daFrame.

mergeClusters(x, table, id)

# S4 method for daFrame
mergeClusters(x, table, id)

Arguments

x
a daFrame.

table
the merging table; a data.frame with columns 'old_cluster', 'new_cluster' and 'label'.

id
character string. Used as a label for the merging.

...
optional arguments.

Value

Writes the newly assigend cluster codes into the metadata slot cluster_codes of the input daFrame and returns the latter.

Details

in the following code snippets, x is a daFrame object.

merging codes are accesible through cluster_codes(x)$id

all functions that ask for specification of a clustering (e.g. plotAbundances, plotClusterHeatmap) take the merging ID as a valid input argument.

References

Nowicka M, Krieg C, Weber LM et al. CyTOF workflow: Differential discovery in high-throughput high-dimensional cytometry datasets. F1000Research 2017, 6:748 (doi: 10.12688/f1000research.11622.1)

Examples

data(PBMC_fs, PBMC_panel, PBMC_md, merging_table)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

#> Error in as.list.default(X): no method for coercing this S4 class to a vector

# run clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33",
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
re <- cluster(re, cols_to_use=lineage)

#> Error in cluster(re, cols_to_use = lineage): object 're' not found

# merge clusters
re <- mergeClusters(re, merging_table, "merging_1")

#> Error in mergeClusters(re, merging_table, "merging_1"): object 're' not found

plotClusterHeatmap(re, k=20, m="merging_1", hm2="pNFkB")

#> Error in plotClusterHeatmap(re, k = 20, m = "merging_1", hm2 = "pNFkB"): object 're' not found