Population-wise flowFrames from a dbFrame

Returns a flowSet or list of flowFrames from a dbFrame. Each flowFrame will contain the subset of events that have been assigned to the same ID.

outFrames(x, ...)

# S4 method for dbFrame
outFrames(x, return = "flowSet", which = "assigned")

Arguments

x
a dbFrame.

...
optional arguments.

return
"flowSet" or "list". Specifies the output type.

which
Specifies which barcode(s) to include. "assigned" (if the population of unassigned events should be excluded), "all" (if the latter should be included), or a numeric or character specifying a subset of populations. Valid values are IDs that occur as row names in the bc_key of the supplied dbFrame. Defaults to "assigned".

Value

a flowSet or list of flowFrames.

Details

Creates a separate flowFrame for each barcode population and, if desired, the population of unassigned events.

Examples

data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
outFrames(x = re, return = "list", which = c("B1", "D4"))

#> [[1]]
#> flowFrame object 'B1'
#> with 601 cells and 6 observables:
#>                name      desc      range    minRange   maxRange
#> (Pd102)Di (Pd102)Di (Pd102)Di 6215.83984 106.6667023 6215.83984
#> (Pd104)Di (Pd104)Di (Pd104)Di   51.80565  -0.9887325   51.80565
#> (Pd105)Di (Pd105)Di (Pd105)Di 5640.69482  93.0029449 5640.69482
#> (Pd106)Di (Pd106)Di (Pd106)Di   54.35886  -0.9903537   54.35886
#> (Pd108)Di (Pd108)Di (Pd108)Di 5215.08789  95.7492523 5215.08789
#> (Pd110)Di (Pd110)Di (Pd110)Di   21.71158  -0.9992667   21.71158
#> 2 keywords are stored in the 'description' slot
#> 
#> [[2]]
#> flowFrame object 'D4'
#> with 751 cells and 6 observables:
#>                name      desc     range   minRange  maxRange
#> (Pd102)Di (Pd102)Di (Pd102)Di  130.5970 -0.9948643  130.5970
#> (Pd104)Di (Pd104)Di (Pd104)Di  120.2310 -0.9946015  120.2310
#> (Pd105)Di (Pd105)Di (Pd105)Di 5746.3184 57.8673897 5746.3184
#> (Pd106)Di (Pd106)Di (Pd106)Di  104.7505 -0.9765739  104.7505
#> (Pd108)Di (Pd108)Di (Pd108)Di 5383.9473 47.5561218 5383.9473
#> (Pd110)Di (Pd110)Di (Pd110)Di 4112.7051 43.5362320 4112.7051
#> 2 keywords are stored in the 'description' slot
#>