Population frequencies across samples & clusters

Plots the relative population abundances of the specified clustering.

plotAbundances(x, ...)

# S4 method for daFrame
plotAbundances(x, k = 20, by = c("sample_id",
  "cluster_id"), group = NULL)

Arguments

x
a daFrame.

...
optional arguments.

k
specifies which clustering to use.

by
a character string specifying whether to plot frequencies by samples or clusters.

group
a character string. Should corresponds to a column name of rowData(x) other than "sample_id" and "cluster_id". The default NULL will use the first factor available.

Value

a ggplot object.

References

Nowicka M, Krieg C, Weber LM et al. CyTOF workflow: Differential discovery in high-throughput high-dimensional cytometry datasets. F1000Research 2017, 6:748 (doi: 10.12688/f1000research.11622.1)

Examples

data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

#> Error in as.list.default(X): no method for coercing this S4 class to a vector

# run clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33",
    "CD123", "CD14", "IgM", "HLA-DR", "CD7")
re <- cluster(re, cols_to_use=lineage)

#> Error in cluster(re, cols_to_use = lineage): object 're' not found

# plot relative population abundances 
plotAbundances(re, k=12)                 # ...by sample 

#> Error in plotAbundances(re, k = 12): object 're' not found

plotAbundances(re, k=8, by="cluster_id") # ...by cluster

#> Error in plotAbundances(re, k = 8, by = "cluster_id"): object 're' not found