Plot cluster heatmap

Plots heatmaps summarizing a clustering and/or metaclustering of interest.

plotClusterHeatmap(x, ...)

# S4 method for daFrame
plotClusterHeatmap(x, hm2 = NULL, k = 20, m = NULL,
  cluster_anno = TRUE, split_by = NULL, scale = TRUE, draw_dend = TRUE,
  draw_freqs = FALSE, palette = rev(brewer.pal(11, "RdYlBu")))

Arguments

x
a daFrame.

...
optional arguments.

hm2
character string. Specifies the right-hand side heatmap. One of:

"abundances": cluster frequencies across samples

"state_markers": median cell state marker expressions across clusters (analogous to the left-hand side heatmap)

a character string/vector corresponding to one/multiple marker(s): median marker expressions across samples and clusters

k
numeric or character string. Specifies the clustering across which median marker expressions should be computed.

m
numeric or character string. Specifies the metaclustering to be shown. (This is for display only and will not effect any computations!)

cluster_anno
logical. Specifies if clusters should be annotated.

split_by
character string. Must corresponds to a column name of rowData(x). If specified, the data will be subset according to this variable, and multiple heatmaps will be drawn.

scale
logical. Specifies whether scaled values should be plotted. (see below for details)

draw_dend
logical. Specifies if the row dendrogram should be drawn.

draw_freqs
logical. Specifyies whether to display cell counts and proportions.

palette
character vector of colors to interpolate.

Value

a HeatmapList-class object.

Details

Scaled values corresponds to cofactor arcsinh-transformed expression values scaled between 0 and 1 using 1 boundaries. Hierarchical clustering is performed on the unscaled data.

In its 1st panel, plotClusterHeatmap will display median (scaled, arcsinh-transformed) cell-type marker expressions (across all samples). Depending on argument hm2, the 2nd panel will contain one of:

relataive cluster abundances by sample

median (scaled, arcsinh-transformed) cell-state marker expressions (across all samples)

median (scaled, arcsinh-transformed) cell-state marker expressions by sample

References

Nowicka M, Krieg C, Weber LM et al. CyTOF workflow: Differential discovery in high-throughput high-dimensional cytometry datasets. F1000Research 2017, 6:748 (doi: 10.12688/f1000research.11622.1)

Examples

data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

#> Error in as.list.default(X): no method for coercing this S4 class to a vector

# run clustering
lineage <- c("CD3", "CD45", "CD4", "CD20", "CD33",
    "CD123", "CD14", "IgM", "HLA_DR", "CD7")
re <- cluster(re, cols_to_use=lineage)

#> Error in cluster(re, cols_to_use = lineage): object 're' not found

plotClusterHeatmap(re, hm2="abundances")

#> Error in plotClusterHeatmap(re, hm2 = "abundances"): object 're' not found

plotClusterHeatmap(re, hm2="state_markers", k=16, split_by='condition')

#> Error in plotClusterHeatmap(re, hm2 = "state_markers", k = 16, split_by = "condition"): object 're' not found

plotClusterHeatmap(re, hm2="pS6", k=12, m=8)

#> Error in plotClusterHeatmap(re, hm2 = "pS6", k = 12, m = 8): object 're' not found