Plot expression heatmap

Plots median marker expressions across samples computed on arcsinh-transformed intensities.

plotExprHeatmap(x, ...)

# S4 method for daFrame
plotExprHeatmap(x, anno = TRUE, color_by = "condition",
  palette = brewer.pal(n = 8, name = "YlGnBu"), scale = TRUE,
  draw_freqs = FALSE, clustering_distance = "euclidean",
  clustering_linkage = "average")

Arguments

x
a daFrame.

...
optional arguments.

anno
logical. Specifies whether to display values insinde each bin.

color_by
character string. Specifies the row annotation.

palette
character vector of colors to interpolate.

scale
logical. Specifies whether to mean-variance normalize each channel.

draw_freqs
logical. Specifyies whether to display cell counts and proportions.

clustering_distance
a character string that specifies the metric to use in dist() for clustering.

clustering_linkage
a character string that specifies the linkage to use in hclust() for clustering.

Value

a HeatmapList-class object.

References

Nowicka M, Krieg C, Weber LM et al. CyTOF workflow: Differential discovery in high-throughput high-dimensional cytometry datasets. F1000Research 2017, 6:748 (doi: 10.12688/f1000research.11622.1)

Examples

data(PBMC_fs, PBMC_panel, PBMC_md)
re <- daFrame(PBMC_fs, PBMC_panel, PBMC_md)

#> Error in as.list.default(X): no method for coercing this S4 class to a vector

plotExprHeatmap(re[, 1:5], draw_freqs=TRUE)

#> Error in plotExprHeatmap(re[, 1:5], draw_freqs = TRUE): object 're' not found