Spillover matrix heat map

Generates a heat map of the spillover matrix annotated with estimated spill percentages.

plotSpillmat(bc_ms, SM, ...)

# S4 method for numeric,matrix
plotSpillmat(bc_ms, SM, out_path = NULL,
  name_ext = NULL, annotate = TRUE, plotly = TRUE,
  isotope_list = CATALYST::isotope_list)

# S4 method for ANY,data.frame
plotSpillmat(bc_ms, SM, out_path = NULL,
  name_ext = NULL, annotate = TRUE, plotly = TRUE)

# S4 method for character,ANY
plotSpillmat(bc_ms, SM, out_path = NULL,
  name_ext = NULL, annotate = TRUE, plotly = TRUE)

Arguments

bc_ms
a vector of numeric masses corresponding to barcode channels.

SM
spillover matrix returned from computeSpillmat.

...
optional arguments.

out_path
character string. If specified, outputs will be generated here.

name_ext
character string. If specified, will be appended to the plot's name.

annotate
logical. If TRUE (default), spill percentages are shown inside bins and rows are annotated with the total amount of spill received.

plotly
logical. Should an interactive plot be rendered?

isotope_list
named list. Used to validate the input spillover matrix. Names should be metals; list elements numeric vectors of their isotopes. See isotope_list for the list of isotopes used by default.

Value

Plots estimated spill percentages as a heat map. Colours are ramped to the highest spillover value present

Examples

# get single-stained control samples
data(ss_exp)

# specify mass channels stained for
bc_ms <- c(139, 141:156, 158:176)

re <- assignPrelim(x = ss_exp, y = bc_ms)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

re <- estCutoffs(x = re)
re <- applyCutoffs(x = re)
spillMat <- computeSpillmat(x = re)
plotSpillmat(bc_ms = bc_ms, SM = spillMat)

#> We recommend that you use the dev version of ggplot2 with `ggplotly()`
#> Install it with: `devtools::install_github('hadley/ggplot2')`