Yield plot

Distribution of barcode separations and yields as a function of separation cutoffs.

plotYields(x, ...)

# S4 method for dbFrame
plotYields(x, which = 0, out_path = NULL,
  name_ext = NULL, plotly = TRUE)

Arguments

x
a dbFrame.

...
optional arguments.

which
0, numeric or character. Specifies which barcode(s) to plot. Valid values are IDs that occur as row names of bc_key(x); 0 (the default) will generate a summary plot with all barcodes.

out_path
character string. If specified, outputs will be generated here.

name_ext
character string. If specified, will be appended to the plot's name.

plotly
logical. Should an interactive plot be rendered?

Value

plots the distribution of barcode separations and yields upon debarcoding as a function of separation cutoffs. If available, currently used separation cutoffs as well as their resulting yields will be indicated in the plot`s main title.

Details

The overall yield that will be achieved upon application of the specified set of separation cutoffs is indicated in the summary plot. Respective separation thresholds and their resulting yields are included in each barcode's plot. The separation cutoff value should be chosen such that it appropriately balances confidence in barcode assignment and cell yield.

References

Zunder, E.R. et al. (2015). Palladium-based mass tag cell barcoding with a doublet-filtering scheme and single-cell deconvolution algorithm. Nature Protocols 10, 316-333.

Examples

data(sample_ff, sample_key)
re <- assignPrelim(x = sample_ff, y = sample_key)

#> Debarcoding data...

#>  o ordering

#>  o classifying events

#> Normalizing...

#> Computing deltas...

#> Computing counts and yields...

re <- estCutoffs(x = re)

# all barcodes summary plot
plotYields(x = re, which = 0)

#> We recommend that you use the dev version of ggplot2 with `ggplotly()`
#> Install it with: `devtools::install_github('hadley/ggplot2')`

#> [[1]]
#> 

# plot for specific sample
plotYields(x = re, which = "C1")

#> We recommend that you use the dev version of ggplot2 with `ggplotly()`
#> Install it with: `devtools::install_github('hadley/ggplot2')`

#> [[1]]
#>